move class to be module
