+++ app/services/json_web_token.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module JsonWebToken
+  module_function
+
+  def encode(payload, exp = 24.hours.from_now)
+    payload[:exp] = exp.to_i
+
+    JWT.encode(payload, Rails.application.secrets.secret_key_base)
+  end
+
+  def decode(token)
+    body = JWT.decode(token, Rails.application.secrets.secret_key_base)[0]
+
+    HashWithIndifferentAccess.new body
+  rescue StandardError
+    nil
+  end
+end
