--- lib/json_web_token.rb
@@ -1,19 +1,0 @@
-# frozen_string_literal: true
-
-class JsonWebToken
-  class << self
-    def encode(payload, exp = 24.hours.from_now)
-      payload[:exp] = exp.to_i
-
-      JWT.encode(payload, Rails.application.secrets.secret_key_base)
-    end
-
-    def decode(token)
-      body = JWT.decode(token, Rails.application.secrets.secret_key_base)[0]
-
-      HashWithIndifferentAccess.new body
-    rescue StandardError
-      nil
-    end
-  end
-end
